+++ tests/test_append_string_no_resize.c
@@ -11,7 +11,7 @@ void test_impl()
   assert_true(string_buffer_append_string(buffer, "abcde"));
   assert_true(string_buffer_append_string(buffer, "ABCDE"));
 
-  assert_num_equal(string_buffer_get_initial_size(buffer), 10240);
+  assert_num_equal(string_buffer_get_initial_size(buffer), STRING_BUFFER_INITIAL_BUFFER_SIZE);
   assert_num_equal(string_buffer_get_content_size(buffer), 15);
   assert_num_equal(string_buffer_get_max_size(buffer), string_buffer_get_initial_size(buffer));